fix: split up token tests
